+++ src/lib/InputInternal.tsx
@@ -22,9 +22,11 @@ const InputInternal = <T extends FieldValues>(props: InputProps<T>) => {
     textAreaRows,
     plainText,
     placeholder,
+    step,
     markAllOnFocus,
     className,
     style,
+    innerRef,
   } = props;
   const { name, id } = useSafeNameId(props.name, props.id);
   const focusHandler = useMarkOnFocusHandler(markAllOnFocus);
@@ -47,7 +49,12 @@ const InputInternal = <T extends FieldValues>(props: InputProps<T>) => {
         invalid={hasError}
         type={type}
         id={id}
-        innerRef={ref}
+        innerRef={(elem) => {
+          if (innerRef) {
+            innerRef.current = elem;
+          }
+          ref(elem);
+        }}
         min={rangeMin}
         max={rangeMax}
         rows={textAreaRows}
@@ -56,6 +63,7 @@ const InputInternal = <T extends FieldValues>(props: InputProps<T>) => {
         plaintext={plainText}
         style={plainText ? { color: "black", marginLeft: 10, ...style } : { ...style }}
         placeholder={placeholder}
+        step={step}
         {...rest}
         {...(value ? { value } : {})}
         onBlur={(e) => {