+++ CHANGELOG.md
@@ -11,6 +11,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `disableBeforeLoading` property (true as default) to `Form` component to prevent the form from being html-native submited by disabling the form till its client-side loaded
 
+### Changed
+
+- `requiredFields` prop type from `(keyof T)[]` to `FieldPath<T>[]`
+
+## [2.6.0] - 2024-02-13
+
+### Changed
+
+- `name` prop type from `keyof T` to `FieldPath<T>`
+
+## [2.5.0] - 2023-12-21
+
+### Added
+
+- Added `innerRef` property to `Input` component
+
+## [2.4.0] - 2023-12-18
+
+### Added
+
+- Added `step` property to `Input` component
+
 ## [2.3.0] - 2023-12-12
 
 ### Added
@@ -354,7 +376,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/2.3.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/2.6.0...HEAD
 [0.1.2]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.1.0
@@ -379,6 +401,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.2.0
+[2.6.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.5.0...2.6.0
+[2.5.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.4.0...2.5.0
+[2.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.3.0...2.4.0
 [2.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.2.2...2.3.0
 [2.2.2]: https://github.com/neolution-ch/react-hook-form-components/compare/2.2.1...2.2.2
 [2.2.1]: https://github.com/neolution-ch/react-hook-form-components/compare/2.2.0...2.2.1